requirements.txt
openai~=0.27.4
flask~=2.2.2
azure-cognitiveservices-speech~=1.27.0
fastapi~=0.95.1
python-multipart~=0.0.6
uvicorn~=0.22.0